requirements.txt
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
youtube-transcript-api>=1.0.0
httpx>=0.27.0
opencc-python-reimplemented>=0.1.7
feedparser>=6.0.0
